INTECH-8

+++ src/main/java/com/example/intech/DAO/MessageDAO.java
@@ -3,5 +3,5 @@ package com.example.intech.DAO;
 import com.example.intech.models.Message;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MessageDAO extends CrudRepository<Message, Integer> {
+public interface MessageDAO extends CrudRepository<Message, Long> {
 }
